# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: XAUUSD Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/notifications/broadcast:
    post:
      operationId: adminBroadcastNotification
      tags: [admin]
      summary: Admin – send a notification to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationBroadcastInput"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppNotification"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/notifications/send-to-users:
    post:
      operationId: adminSendToUsers
      tags: [admin]
      summary: Admin – send notification to specific users only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationTargetedInput"
      responses:
        "201":
          description: Notifications sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent:
                    type: integer
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/notifications/{id}:
    delete:
      operationId: adminDeleteNotification
      tags: [admin]
      summary: Admin – delete a single broadcast notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          description: Not found
  /admin/system-settings:
    get:
      operationId: adminGetSystemSettings
      tags: [admin]
      summary: Admin – get system settings (licenseEnforcementEnabled, trialModeEnabled)
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    put:
      operationId: adminUpdateSystemSettings
      tags: [admin]
      summary: Admin – update system settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AppNotification:
      type: object
      required: [id, userId, type, title, body, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          description: "daily | weekly | monthly | win_alert | loss_alert | off_day | admin"
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    NotificationBroadcastInput:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
        body:
          type: string
    NotificationTargetedInput:
      type: object
      required: [title, body, userIds]
      properties:
        title:
          type: string
        body:
          type: string
        userIds:
          type: array
          items:
            type: string
          description: Clerk user IDs to send this notification to
...
    SystemSettings:
      type: object
      required: [licenseEnforcementEnabled, trialModeEnabled, trialDurationDays]
      properties:
        licenseEnforcementEnabled:
          type: boolean
          description: "When false, all users can access the terminal without a license"
        trialModeEnabled:
          type: boolean
          description: "When true, users can self-activate a free trial"
        trialDurationDays:
          type: integer
          description: "Number of days for the free trial (admin-configurable)"
    SystemSettingsInput:
      type: object
      properties:
        licenseEnforcementEnabled:
          type: boolean
        trialModeEnabled:
          type: boolean
        trialDurationDays:
          type: integer

    # ACTIVITY LOG
...